Implementation of the TCF v2 consent string spec
This change implements the newly released TCF v2 consent string spec.
More can be found on this here:

https://github.com/InteractiveAdvertisingBureau/GDPR-Transparency-and-Consent-Framework/blob/master/TCFv2

Due to backward compatability issues, we have multiple similarly named
types, and functions. Until v1.1 is no longer expected in the wild, we
will continue to support this. We may later decide to create a v2 module
within the repo, or at a new location.

I'm honestly still not quite confident on how to implement the methods
EveryPurposeAllowed and VendorAllowed for the new type V2ParsedConsent,
and that may need to come in a subsequent PR.