Implementation of the TCF v2 consent string spec (#11)
This change implements the newly released TCF v2 consent string spec.
More can be found on this here:

https://github.com/InteractiveAdvertisingBureau/GDPR-Transparency-and-Consent-Framework/blob/master/TCFv2

Due to backward compatibility issues, we have multiple similarly named
types, and functions. Until v1.1 is no longer expected in the wild, we
will continue to support this. We may later decide to create a v2 module
within the repo, or at a new location.

We did not attempt to implement EveryPurposeAllowed and VendorAllowed
for the new type V2ParsedConsent, and that should come in a subsequent
change. Also ideally we'll add some additional unit test coverage as well.

Reviewers: rupertchen, imang1