Add more v2 processing logic, tests abound (#15)
Given discussions with our data ethics team we have come up with what we think the proper logic is for determining whether it is legal to process data in both v1 and v2 based on the bases of consented (or opted-in) purposes. It's worth noting that the v2 library does NOT have any built in logic for determining whether it's legal to process on the basis of legitimate interest (as LiveRamp will not be doing this). We may chose to add this in later (or would be open to a PR to support this).

Reviewers: mdedhi, julien-weber